requirements.txt
dt_apriltags==3.1.7
matplotlib==3.10.7
numpy==2.2.6
opencv_python==4.12.0.88
pyrealsense2==2.56.5.9235
spatialmath_python==1.1.14
xarm-python-sdk==1.17.0
